:fire: Restructure files and folders
Delete unnecessary starter files
