--- client/src/components/Header/Header.stories.js
@@ -1,18 +1,0 @@
-// import React from "react";
-
-// import Header from "./Header";
-
-// export default {
-//   title: "Example/Header",
-//   component: Header,
-// };
-
-// const Template = (args) => <Header {...args} />;
-
-// export const LoggedIn = Template.bind({});
-// LoggedIn.args = {
-//   user: {},
-// };
-
-// export const LoggedOut = Template.bind({});
-// LoggedOut.args = {};
