--- client/src/components/Header/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Header";
