+++ client/src/components/IconButton.js
@@ -1,13 +1,13 @@
 import styled from "styled-components/macro";
 import PropTypes from "prop-types";
-import { ReactComponent as ThemeToggler } from "../../assets/icon-sun-empty.svg";
-import { ReactComponent as BurgerMenu } from "../../assets/icon-burger-menu.svg";
-import { ReactComponent as Close } from "../../assets/icon-close.svg";
-import { ReactComponent as ArrowBack } from "../../assets/icon-arrow-back.svg";
-import { ReactComponent as LogOut } from "../../assets/icon-logout.svg";
-import { ReactComponent as Filter } from "../../assets/icon-filter.svg";
-import { ReactComponent as Info } from "../../assets/icon-info-full.svg";
-import { ReactComponent as AddToSpotify } from "../../assets/add-to-spotify.svg";
+import { ReactComponent as ThemeToggler } from "../assets/icon-sun-empty.svg";
+import { ReactComponent as BurgerMenu } from "../assets/icon-burger-menu.svg";
+import { ReactComponent as Close } from "../assets/icon-close.svg";
+import { ReactComponent as ArrowBack } from "../assets/icon-arrow-back.svg";
+import { ReactComponent as LogOut } from "../assets/icon-logout.svg";
+import { ReactComponent as Filter } from "../assets/icon-filter.svg";
+import { ReactComponent as Info } from "../assets/icon-info-full.svg";
+import { ReactComponent as AddToSpotify } from "../assets/add-to-spotify.svg";
 
 export const IconButton = styled.button`
   display: flex;
