--- client/src/pages/Page/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Page";
